refactor: extract shared widgets and component catalog
- Add SectionHeader widget replacing private _SectionHeader in 2 screens
- Add AppNavigationTile widget replacing private _HelpTile, _SettingsTile,
  _LegalTile across 4 screens
- Add AdaptiveColorCalculator utility consolidating duplicated luminance
  logic from LocationChip and ExpandableLocationPanel
- Fix ExpandableLocationPanel using ad-hoc hex colors instead of
  BrandPalette tokens (consistency bug)
- Add component-catalog.instructions.md for Copilot auto-attach on UI files

Screens migrated: help_info_screen, settings_screen, about_screen,
about_settings_screen
Tests: 1867 pass, 28 skip, 7 pre-existing golden failures (unchanged)
