+++ lib/features/settings/screens/about_settings_screen.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
+import 'package:wildfire_mvp_v3/widgets/app_navigation_tile.dart';
+
 /// About section within Settings showing legal and policy documents.
 ///
 /// This provides a dedicated screen for legal documents, though they
@@ -33,28 +35,28 @@ class AboutSettingsScreen extends StatelessWidget {
               ),
             ),
 
-            _LegalTile(
+            AppNavigationTile(
               icon: Icons.description_outlined,
               title: 'Terms of Service',
               subtitle: 'App usage terms and conditions',
               onTap: () => context.push('/settings/about/terms'),
             ),
 
-            _LegalTile(
+            AppNavigationTile(
               icon: Icons.privacy_tip_outlined,
               title: 'Privacy Policy',
               subtitle: 'How we collect and use your data',
               onTap: () => context.push('/settings/about/privacy'),
             ),
 
-            _LegalTile(
+            AppNavigationTile(
               icon: Icons.warning_amber_outlined,
               title: 'Emergency Disclaimer',
               subtitle: 'Important safety information and limitations',
               onTap: () => context.push('/settings/about/disclaimer'),
             ),
 
-            _LegalTile(
+            AppNavigationTile(
               icon: Icons.source_outlined,
               title: 'Data Sources',
               subtitle: 'Data providers and attribution',
@@ -78,42 +80,6 @@ class AboutSettingsScreen extends StatelessWidget {
           ],
         ),
       ),
-    );
-  }
-}
-
-/// A list tile for legal document navigation.
-class _LegalTile extends StatelessWidget {
-  final IconData icon;
-  final String title;
-  final String subtitle;
-  final VoidCallback onTap;
-
-  const _LegalTile({
-    required this.icon,
-    required this.title,
-    required this.subtitle,
-    required this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-
-    return ListTile(
-      leading: Icon(icon, color: theme.colorScheme.onSurfaceVariant),
-      title: Text(title),
-      subtitle: Text(
-        subtitle,
-        style: theme.textTheme.bodySmall?.copyWith(
-          color: theme.colorScheme.onSurfaceVariant,
-        ),
-      ),
-      trailing: Icon(
-        Icons.chevron_right,
-        color: theme.colorScheme.onSurfaceVariant,
-      ),
-      onTap: onTap,
     );
   }
 }
